updates on the schedule-groups and schedule-epochs

+++ dashboards/xbosdash/static/js/schedule/schedule-groupings.js
@@ -14,15 +14,19 @@ $(document).ready(function() {
 
 	function getModes() {
 		$.ajax({
-			"url": "http://0.0.0.0:5000/get_modes",
+			"url": "http://0.0.0.0:5000/api/get_modes",
 			"type": "GET",
 			"dataType": "json",
 			"success": function(d) {
-				console.log("hi");
-			},
+				// if (d["modes"] == null) { d = [{id: 0, name: "Closed", heating: "55", cooling: "85", enabled: true}, {id: 1, name: "Open", heating: "70", cooling: "75", enabled: false}, {id: 2, name: "Do Not Exceed", heating: "52", cooling: "83", enabled: true}, {id: 3, name: "Other", heating: "51", cooling: "84", enabled: true}, {id: 4, name: "Midnight", heating: "54", cooling: "88", enabled: false}, {id: 5, name: "Early Morn", heating: "50", cooling: "80", enabled: true}]; }
+				console.log('success get_modes from mongodb: ', d)
+				modeSettings = d['success'][0]['data'].sort(mySort);
+				localStorage.setItem("mode-settings", JSON.stringify(modeSettings));
+				setModes();			},
 			"error": function(d) {
-				if (d["modes"] == null) { d = {modes: [{id: 0, name: "Closed", heating: "55", cooling: "85", enabled: true}, {id: 1, name: "Open", heating: "70", cooling: "75", enabled: false}, {id: 2, name: "Do Not Exceed", heating: "52", cooling: "83", enabled: true}, {id: 3, name: "Other", heating: "51", cooling: "84", enabled: true}, {id: 4, name: "Midnight", heating: "54", cooling: "88", enabled: false}, {id: 5, name: "Early Morn", heating: "50", cooling: "80", enabled: true}]}; }
-				modeSettings = d["modes"].sort(mySort);
+				// if (d["modes"] == null) { d = [{id: 0, name: "Closed", heating: "55", cooling: "85", enabled: true}, {id: 1, name: "Open", heating: "70", cooling: "75", enabled: false}, {id: 2, name: "Do Not Exceed", heating: "52", cooling: "83", enabled: true}, {id: 3, name: "Other", heating: "51", cooling: "84", enabled: true}, {id: 4, name: "Midnight", heating: "54", cooling: "88", enabled: false}, {id: 5, name: "Early Morn", heating: "50", cooling: "80", enabled: true}]; }
+				console.log('error get_modes from mongodb: ', d)
+				modeSettings = d.sort(mySort);
 				localStorage.setItem("mode-settings", JSON.stringify(modeSettings));
 				setModes();
 			}
@@ -31,6 +35,7 @@ $(document).ready(function() {
 
 	function setModes() {
 		var modeDiv = $("#mode-div");
+		modeDiv.html("");
 		var count = 0;
 		for (var i in modeSettings) {
 			var curr = modeSettings[i];
@@ -44,22 +49,23 @@ $(document).ready(function() {
 
 	function getGroupings() {
 		$.ajax({
-			"url": "http://0.0.0.0:5000/get_groups",
+			"url": "http://0.0.0.0:5000/api/get_groupings",
 			"type": "GET",
 			"dataType": "json",
 			"success": function(d) {
-				console.log("hi");
+				console.log('/api/get_groupings success: ', d["success"]);
+				allGroupings = d["success"];
+				localStorage.setItem("all-groupings", JSON.stringify(allGroupings));
+				setGroupings();
 			},
 			"error": function(d) {
-				if (d["groupings"] == null) { d = {groupings:[{group: "Basketball Courts 1", zones: ["North 1", "South 1", "East 1", "West 1"],settings:{fri: [2, 1, 0], mon: [2, 1, 0], sat: [4, 3, 2], sun: [0, 2, 1], thu: [1, 3, 4], tue: [4, 3, 4], wed: [1, 0, 2]},times: {fri: ["8.00", "18.00"], mon: ["2.00", "17.00"], sat: ["3.00", "16.00"], sun: ["4.00", "15.00"], thu: ["5.00", "14.00"], tue: ["6.00", "13.00"], wed: ["7.00", "12.00"]}},{group: "Basketball Courts 2", zones: ["North 2", "South 2", "East 2", "West 2"],settings:{sun: [2, 1, 0], mon: [2, 1, 0], thu: [4, 3, 2], fri: [0, 2, 1], sat: [1, 3, 4], wed: [4, 3, 4], tue: [1, 0, 2]},times: {fri: ["8.00", "18.00"], mon: ["2.00", "17.00"], sat: ["3.00", "16.00"], sun: ["4.00", "15.00"], thu: ["5.00", "14.00"], tue: ["6.00", "13.00"], wed: ["7.00", "12.00"]}},{group: "Basketball Courts 3", zones: ["North 3", "South 3", "East 3", "West 3"],settings:{mon: [2, 1, 0], fri: [2, 1, 0], sat: [4, 3, 2], thu: [0, 2, 1], sun: [1, 3, 4], tue: [4, 3, 4], wed: [1, 0, 2]},times: {fri: ["8.00", "18.00"], mon: ["2.00", "17.00"], sat: ["3.00", "16.00"], sun: ["4.00", "15.00"], thu: ["5.00", "14.00"], tue: ["6.00", "13.00"], wed: ["7.00", "12.00"]}},{group: "Basketball Courts 4", zones: ["North 4", "South 4", "East 4", "West 4"],settings:{fri: [2, 1, 0], sat: [2, 1, 0], mon: [4, 3, 2], sun: [0, 2, 1], tue: [1, 3, 4], thu: [4, 3, 4], wed: [1, 0, 2]},times: {fri: ["8.00", "18.00"], mon: ["2.00", "17.00"], sat: ["3.00", "16.00"], sun: ["4.00", "15.00"], thu: ["5.00", "14.00"], tue: ["6.00", "13.00"], wed: ["7.00", "12.00"]}}]}; }
-				allGroupings = d["groupings"];
+				if (d["groupings"] == null) { d = [{group: "Basketball Courts 1", settings:{fri: [2, 1, 0], mon: [2, 1, 0], sat: [4, 3, 2], sun: [0, 2, 1], thu: [1, 3, 4], tue: [4, 3, 4], wed: [1, 0, 2]},times: {fri: ["8.00", "18.00"], mon: ["2.00", "17.00"], sat: ["3.00", "16.00"], sun: ["4.00", "15.00"], thu: ["5.00", "14.00"], tue: ["6.00", "13.00"], wed: ["7.00", "12.00"]}, zones: ["North 1", "South 1", "East 1", "West 1"]},{group: "Basketball Courts 2", settings:{sun: [2, 1, 0], mon: [2, 1, 0], thu: [4, 3, 2], fri: [0, 2, 1], sat: [1, 3, 4], wed: [4, 3, 4], tue: [1, 0, 2]},times: {fri: ["8.00", "18.00"], mon: ["2.00", "17.00"], sat: ["3.00", "16.00"], sun: ["4.00", "15.00"], thu: ["5.00", "14.00"], tue: ["6.00", "13.00"], wed: ["7.00", "12.00"]}, zones: ["North 2", "South 2", "East 2", "West 2"]},{group: "Basketball Courts 3", settings:{mon: [2, 1, 0], fri: [2, 1, 0], sat: [4, 3, 2], thu: [0, 2, 1], sun: [1, 3, 4], tue: [4, 3, 4], wed: [1, 0, 2]},times: {fri: ["8.00", "18.00"], mon: ["2.00", "17.00"], sat: ["3.00", "16.00"], sun: ["4.00", "15.00"], thu: ["5.00", "14.00"], tue: ["6.00", "13.00"], wed: ["7.00", "12.00"]}, zones: ["North 3", "South 3", "East 3", "West 3"]},{group: "Basketball Courts 4", settings:{fri: [2, 1, 0], sat: [2, 1, 0], mon: [4, 3, 2], sun: [0, 2, 1], tue: [1, 3, 4], thu: [4, 3, 4], wed: [1, 0, 2]},times: {fri: ["8.00", "18.00"], mon: ["2.00", "17.00"], sat: ["3.00", "16.00"], sun: ["4.00", "15.00"], thu: ["5.00", "14.00"], tue: ["6.00", "13.00"], wed: ["7.00", "12.00"]}, zones: ["North 4", "South 4", "East 4", "West 4"]}]; }
+				allGroupings = d;
 				localStorage.setItem("all-groupings", JSON.stringify(allGroupings));
 				setGroupings();
 			}
 		});
 	}
-
-	var groupNames = [];
 
 	function setGroupings() {
 		if (allGroupings.length == 0) { return; }
@@ -68,9 +74,8 @@ $(document).ready(function() {
 		for (var i in allGroupings) {
 			var curr = allGroupings[i];
 			var name = curr["group"];
-			groupNames.push(name);
 			localStorage.setItem(name + "-group", JSON.stringify(curr));
-			groupings.append("<div class='col s3-3 group-card'><div class='card hoverable'><div class='card-content'><!-- <p>Current Heating Setpoint: 65<br>Current Cooling Setpoint: 78<br>Current State: On</p> --><span class='card-title' style='margin-bottom: 0; margin-top: 24px;'>" + name + "</span></div><div class='card-action'><a class='ca-modes' href='#'>Modes</a><!-- <a class='ca-zones' href='#'>Zones</a> --><a class='ca-delete' href='#'>Delete</a></div></div></div>");
+			groupings.append("<div class='col s3-3 group-card'><div class='card hoverable'><div class='card-content'><!-- <p>Current Heating Setpoint: 65<br>Current Cooling Setpoint: 78<br>Current State: On</p> --><span class='card-title' style='margin-bottom: 0; margin-top: 24px;'>" + name + "</span></div><div class='card-action'><a class='ca-modes' href='#'>Edit</a><a class='ca-zones modal-trigger' href='#modal1'>Zones</a><a class='ca-delete' href='#'>Delete</a></div></div></div>");
 			$.merge(usedZones, curr["zones"]);
 		}
 		setGroupingClicks();
@@ -78,39 +83,66 @@ $(document).ready(function() {
 
 	function setGroupingClicks() {
 		$(".ca-modes").each(function(i) {
-			var name = groupNames[i];
+			var name = allGroupings[i]["group"];
 			$(this).click(function() {
 				localStorage.setItem("curr-group-name", name);
 				localStorage.removeItem("zones-for-group");
 				window.location.href = "schedule-epochs.html";
 			});
 		});
-		// $(".ca-zones").each(function() {
-		// 	$(this).click(function() {
-		// 		console.log($(this));
-		// 	});
-		// });
-		$(".ca-delete").each(function(i) {
-			var name = groupNames[i];
+		$(".ca-zones").each(function(i) {
+			var name = allGroupings[i]["group"];
+			var zones = allGroupings[i]["zones"];
 			$(this).click(function() {
-				deleteGrouping(name);
+				$("#m-footer").hide();
+				var s = "'" + name + "' contains ";
+				if (zones.length == 1) { s += "1 zone:"; }
+				else { s += zones.length + " zones:"; }
+				$("#modal-header").html(s);
+				$("#modal-text").html(zones.join("<br>"));
+			});
+		});
+		$(".ca-delete").each(function(i) {
+			var name = allGroupings[i]["group"];
+			$(this).click(function() {
+				if (confirm("Are you sure you want to delete " + name + "?")) {
+					deleteGrouping(name);
+				}
 			});
 		});
 	}
-
+	
 	if (allGroupings == null) { getGroupings(); } else { setGroupings(); }
 
+	function findGroup(name) {
+		for (var i in allGroupings) {
+			if (allGroupings[i]["group"] == name) {
+				return i;
+			}
+		}
+	}
+
+	// http://www.jquerybyexample.net/2012/02/remove-item-from-array-using-jquery.html
 	function deleteGrouping(name) {
 		localStorage.removeItem(name + "-group");
+		allGroupings.splice(findGroup(name), 1);
+		localStorage.setItem("all-groupings", JSON.stringify(allGroupings));
+
+		// Reload causes the entire console history to be deleted too.
+		window.location.reload();
+
+		safeToast("Grouping successfully deleted.", "rounded");
 		$.ajax({
-			"url": "http://0.0.0.0:5000/delete_grouping",
-			"type": "GET",
+			"url": "http://0.0.0.0:5000/api/delete_grouping",
+			"type": "POST",
 			"dataType": "json",
+			"contentType": 'application/json',
+			"data": JSON.stringify({'group': name}),
 			"success": function(d) {
-				console.log("hi");
+				console.log("/api/delete_grouping success: ", d);
 			},
 			"error": function(d) {
-				console.log(name);
+				console.log('/api/delete_grouping error: ', name);
 			}
 		});
 	}
@@ -123,14 +155,20 @@ $(document).ready(function() {
 
 	function getCheckboxes() {
 		$.ajax({
-			"url": "http://0.0.0.0:5000/get_zones",
+			"url": "http://0.0.0.0:5000/api/get_zones",
 			"type": "GET",
 			"dataType": "json",
 			"success": function(d) {
-				console.log("hi");
+				console.log("/api/get_zones success: ", d);
+				allZones = d["success"].split(",");
+				localStorage.setItem("all-zones", allZones);
+				setCheckboxes();
 			},
 			"error": function(d) {
-				if (d["zones"] == null) { d = {zones: ["North 1", "South 1", "East 1", "West 1", "North 2", "South 2", "East 2", "West 2 Basketball", "North 3 Basketball", "South 3 Basketball", "East 3 Basketball", "West 3 Basketball", "North 4 Basketball", "South 4 Basketball", "East 4 Basketball", "West 4 Basketball", "North 5 Basketball", "North 6 Basketball", "South 5 Basketball", "South 6", "East 5", "East 6", "West 5", "West 6"]}; }
+			    console.log("/api/get_zones error: ", d)
+				if (d["zones"] == null) {
+				    d = { zones: ["North 1", "South 1", "East 1", "West 1", "North 2", "South 2", "East 2", "West 2", "North 3", "South 3", "East 3", "West 3", "North 4", "South 4", "East 4", "West 4", "North 5 Basketball", "North 6 Basketball", "South 5 Basketball", "South 6", "East 5", "East 6", "West 5", "West 6"]};
+				}
 				allZones = d["zones"];
 				localStorage.setItem("all-zones", JSON.stringify(allZones));
 				setCheckboxes(d);
@@ -182,6 +220,7 @@ $(document).ready(function() {
 	$("#group-btn").click(function() {
 		var s = "Form a new group with the following ";
 		if (zoneSel == 1) { s += "zone:"; } else { s += zoneSel + " zones:"; }
+		$("#m-footer").show();
 		$("#modal-header").html(s);
 		$("#modal-text").html(zoneArr.join("<br>"));
 	});
@@ -205,8 +244,7 @@ $(document).ready(function() {
 	$("#save-modes").click(function() { smartSaveModes() } );
 
 	function createModeObj() {
-		var obj = new Object();
-		obj.modes = [];
+		var modes = [];
 		$(".mode-card").each(function(i) {
 			var m = new Object();
 			m.id = i;
@@ -215,9 +253,9 @@ $(document).ready(function() {
 			m.heating = inputs[1].value;
 			m.cooling = inputs[2].value;
 			m.enabled = $(inputs[3]).prop("checked");
-			obj.modes.push(m);
+			modes.push(m);
 		});
-		return obj;
+		return modes;
 	}
 
 	function safeToast(s, c, t=2000) {
@@ -228,22 +266,25 @@ $(document).ready(function() {
 	function smartSaveModes() {
 		var obj = createModeObj();
 		var stringified = JSON.stringify(obj);
+		console.log('stringified: ', stringified)
 
 		if (stringified == localStorage.getItem("mode-settings")) {
 			safeToast("No changes to be saved.", "rounded");
 		} else {
 			localStorage.setItem("mode-settings", stringified);
 			$.ajax({
-				"url": "http://0.0.0.0:5000/save_modes",
-				"type": "GET",
+				"url": "http://0.0.0.0:5000/api/save_modes",
+				"type": "POST",
 				"dataType": "json",
+				"contentType": 'application/json',
+				"data": stringified,
 				"success": function(d) {
-					console.log("hi");
+				    console.log('/api/save_modes success: ', d);
 				},
 				"error": function(d) {
 					safeToast("Current modes successfully updated.", "rounded");
-					console.log(obj);
-					console.log(stringified);
+					// console.log(obj);
+					// console.log(stringified);
 				}
 			});	
 		}
